perf(native, binding): move Casting utils to binding module, add TODO
Casting to custom classes requires getting a kotlin instance instead of creating a new one
